db: move drizzle schema from worker/view/drizzle to worker/db/drizzle
D1 is authoritative for every product domain, not a "view" derived from
elsewhere. The directory move reflects that: schema, migrations, and
drizzle.config.ts all relocate under worker/db/. Table shapes are
unchanged.

- wrangler.jsonc migrations_dir points at the new path
- 5 reader modules and 17 integration tests update their import paths
- drizzle.config.ts schema + out paths follow the move

First structural step of ADR 0009 (d1-direct).
