+++ apps/web/tests/integration/pano-post-connection.test.ts
@@ -16,10 +16,10 @@ import {env} from "cloudflare:test";
 import {id} from "@usirin/forge";
 import {beforeAll, describe, expect, it} from "vitest";
 import {listPostConnection} from "../../worker/features/pano/postSummaryReader";
-import viewMigration0000 from "../../worker/view/drizzle/migrations/0000_secret_iron_patriot.sql";
-import viewMigration0001 from "../../worker/view/drizzle/migrations/0001_free_salo.sql";
-import viewMigration0002 from "../../worker/view/drizzle/migrations/0002_wandering_natasha_romanoff.sql";
-import viewMigration0003 from "../../worker/view/drizzle/migrations/0003_lazy_thanos.sql";
+import viewMigration0000 from "../../worker/db/drizzle/migrations/0000_secret_iron_patriot.sql";
+import viewMigration0001 from "../../worker/db/drizzle/migrations/0001_free_salo.sql";
+import viewMigration0002 from "../../worker/db/drizzle/migrations/0002_wandering_natasha_romanoff.sql";
+import viewMigration0003 from "../../worker/db/drizzle/migrations/0003_lazy_thanos.sql";
 
 declare module "cloudflare:test" {
 	// biome-ignore lint/suspicious/noEmptyBlockStatements: required by pool-workers
