+++ apps/web/worker/db/drizzle.config.ts
@@ -3,6 +3,6 @@ import {defineConfig} from "drizzle-kit";
 export default defineConfig({
 	dialect: "sqlite",
 	driver: "d1-http",
-	schema: "./worker/view/drizzle/schema.ts",
-	out: "./worker/view/drizzle/migrations",
+	schema: "./worker/db/drizzle/schema.ts",
+	out: "./worker/db/drizzle/migrations",
 });
