+++ apps/web/worker/features/pasaport/userProfileReader.ts
@@ -17,7 +17,7 @@
  */
 import {and, desc, eq, isNull, lt, or, sql} from "drizzle-orm";
 import {drizzle} from "drizzle-orm/d1";
-import * as schema from "../../view/drizzle/schema";
+import * as schema from "../../db/drizzle/schema";
 
 /* -------------------------------------------------------------------------- */
 /* Identity + aggregates                                                       */
