+++ apps/web/worker/features/sozluk/termSummaryReader.ts
@@ -10,7 +10,7 @@
  */
 import {and, desc, eq, gt, lt, or, sql} from "drizzle-orm";
 import {drizzle} from "drizzle-orm/d1";
-import * as schema from "../../view/drizzle/schema";
+import * as schema from "../../db/drizzle/schema";
 
 export type ListSort = "recent" | "popular";
